api/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
minio==7.2.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0
Pillow==10.1.0
Werkzeug==3.0.1
